backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==12.0.0
